docs: migrate results component
