+++ packages/atomic/src/components/search/result-template-components/atomic-field-condition/atomic-field-condition.new.stories.tsx
@@ -1,0 +1,62 @@
+import {renderComponent} from '@coveo/atomic/storybookUtils/render-component';
+import {wrapInResult} from '@coveo/atomic/storybookUtils/result-wrapper';
+import {wrapInSearchInterface} from '@coveo/atomic/storybookUtils/search-interface-wrapper';
+import type {Meta, StoryObj} from '@storybook/web-components';
+import {html} from 'lit/static-html.js';
+
+const {decorator: resultDecorator, engineConfig} = wrapInResult();
+const {decorator: searchInterfaceDecorator, play} =
+  wrapInSearchInterface(engineConfig);
+
+const meta: Meta = {
+  component: 'atomic-field-condition',
+  title: 'Atomic/ResultList/FieldCondition',
+  id: 'atomic-field-condition',
+
+  render: renderComponent,
+  decorators: [resultDecorator, searchInterfaceDecorator],
+  parameters: {
+    controls: {expanded: true, hideNoControlsWarning: true},
+  },
+  play,
+};
+
+export default meta;
+type Story = StoryObj;
+
+export const Default: Story = {
+  name: 'atomic-field-condition',
+  args: {
+    default: `
+      <div>
+        The visibility of this text can be controlled by the field conditions component
+      </div>
+    `,
+  },
+  decorators: [
+    (story) => html`
+      ${story()}
+      <div style="margin:20px 0">
+        Select facet value(s) to filter on field values and influence the
+        condition(s).
+      </div>
+      <div style="display: flex; justify-content: flex-start;">
+        <atomic-facet
+          field="objecttype"
+          style="flex-grow:1"
+          label="Object type"
+        ></atomic-facet>
+        <atomic-facet
+          field="filetype"
+          style="flex-grow:1"
+          label="File type"
+        ></atomic-facet>
+        <atomic-facet
+          field="source"
+          style="flex-grow:1"
+          label="Source"
+        ></atomic-facet>
+      </div>
+    `,
+  ],
+};
