+++ packages/atomic/src/components/search/result-template-components/atomic-quickview/atomic-quickview.new.stories.tsx
@@ -1,0 +1,53 @@
+import {renderComponent} from '@coveo/atomic/storybookUtils/render-component';
+import {wrapInResult} from '@coveo/atomic/storybookUtils/result-wrapper';
+import {wrapInSearchInterface} from '@coveo/atomic/storybookUtils/search-interface-wrapper';
+import type {Meta, StoryObj} from '@storybook/web-components';
+
+const {decorator: resultDecorator, engineConfig} = wrapInResult({
+  preprocessRequest: (request) => {
+    const fakeQuery = '';
+    switch (
+      (request.headers as Record<string, string> | undefined)?.['Content-Type']
+    ) {
+      case 'application/json': {
+        const parsed = JSON.parse(request.body as string);
+        parsed.aq = '@filetype=pdf';
+        parsed.q = fakeQuery;
+        parsed.numberOfResults = 1;
+        request.body = JSON.stringify(parsed);
+        break;
+      }
+      case 'application/x-www-form-urlencoded':
+        if (!request.body) {
+          break;
+        }
+        request.body = (request.body as string).replace(
+          'q=&',
+          `q=${encodeURIComponent(fakeQuery)}&`
+        );
+        break;
+    }
+    return request;
+  },
+});
+const {decorator: searchInterfaceDecorator, play} =
+  wrapInSearchInterface(engineConfig);
+
+const meta: Meta = {
+  component: 'atomic-quickview',
+  title: 'Atomic/ResultList/Quickview',
+  id: 'atomic-quickview',
+  render: renderComponent,
+  decorators: [resultDecorator, searchInterfaceDecorator],
+  parameters: {
+    controls: {expanded: true, hideNoControlsWarning: true},
+  },
+  play,
+};
+
+export default meta;
+type Story = StoryObj;
+
+export const Default: Story = {
+  name: 'atomic-quick-view',
+};
