--- packages/atomic/src/components/search/result-template-components/atomic-result-badge/atomic-result-badge.stories.tsx
@@ -1,13 +1,0 @@
-import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'atomic-result-badge',
-  {field: 'filetype'}
-);
-
-export default {
-  ...defaultModuleExport,
-  title: 'Atomic/ResultList/ResultBadge',
-  id: 'atomic-result-badge',
-};
-export const Default = exportedStory;
