--- packages/atomic/src/components/search/result-template-components/atomic-result-date/atomic-result-date.stories.tsx
@@ -1,12 +1,0 @@
-import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'atomic-result-date',
-  {}
-);
-export default {
-  ...defaultModuleExport,
-  title: 'Atomic/ResultList/ResultDate',
-  id: 'atomic-result-date',
-};
-export const Default = exportedStory;
