+++ packages/atomic/src/components/search/result-template-components/atomic-result-fields-list/atomic-result-fields-list.tsx
@@ -4,6 +4,7 @@ import {ResultContext} from '../result-template-decorators';
 
 /**
  * The `atomic-result-fields-list` component selectively renders its children to ensure they fit the parent element and adds dividers between them.
+ * @slot default - The children to render.
  */
 @Component({
   tag: 'atomic-result-fields-list',
