--- packages/atomic/src/components/search/result-template-components/atomic-result-icon/atomic-result-icon.stories.tsx
@@ -1,13 +1,0 @@
-import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'atomic-result-icon',
-  {}
-);
-
-export default {
-  ...defaultModuleExport,
-  title: 'Atomic/ResultList/ResultIcon',
-  id: 'atomic-result-icon',
-};
-export const Default = exportedStory;
