+++ packages/atomic/src/components/search/result-template-components/atomic-result-link/atomic-result-link.new.stories.tsx
@@ -1,16 +1,18 @@
 import {renderComponent} from '@coveo/atomic/storybookUtils/render-component';
+import {wrapInResult} from '@coveo/atomic/storybookUtils/result-wrapper';
 import {wrapInSearchInterface} from '@coveo/atomic/storybookUtils/search-interface-wrapper';
 import type {Meta, StoryObj} from '@storybook/web-components';
 
-const {decorator, play} = wrapInSearchInterface();
+const {decorator: resultDecorator, engineConfig} = wrapInResult();
+const {decorator: searchInterfaceDecorator, play} =
+  wrapInSearchInterface(engineConfig);
 
 const meta: Meta = {
-  component: 'atomic-load-more-results',
-  title: 'Atomic/LoadMoreResults',
-  id: 'atomic-load-more-results',
-
+  component: 'atomic-result-icon',
+  title: 'Atomic/ResultList/ResultLink',
+  id: 'atomic-result-link',
   render: renderComponent,
-  decorators: [decorator],
+  decorators: [resultDecorator, searchInterfaceDecorator],
   parameters: {
     controls: {expanded: true, hideNoControlsWarning: true},
   },
@@ -21,6 +23,5 @@ export default meta;
 type Story = StoryObj;
 
 export const Default: Story = {
-  name: 'atomic-load-more-results',
-  args: {},
+  name: 'atomic-result-icon',
 };
