--- packages/atomic/src/components/search/result-template-components/atomic-result-link/atomic-result-link.stories.tsx
@@ -1,13 +1,0 @@
-import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'atomic-result-link',
-  {}
-);
-
-export default {
-  ...defaultModuleExport,
-  title: 'Atomic/ResultList/ResultLink',
-  id: 'atomic-result-link',
-};
-export const Default = exportedStory;
