+++ packages/atomic/src/components/search/result-template-components/atomic-result-number/atomic-result-number.new.stories.tsx
@@ -4,22 +4,21 @@ import {wrapInSearchInterface} from '@coveo/atomic/storybookUtils/search-interfa
 import type {Meta, StoryObj} from '@storybook/web-components';
 
 const {decorator: resultDecorator, engineConfig} = wrapInResult({
-  search: {
-    preprocessSearchResponseMiddleware: (res) => {
-      res.body.results.forEach(
-        (r) => (r.raw['randomimage'] = 'https://picsum.photos/200')
-      );
-      return res;
-    },
+  preprocessRequest: (r) => {
+    const parsed = JSON.parse(r.body as string);
+    parsed.fieldsToInclude = [...parsed.fieldsToInclude, 'size'];
+    parsed.numberOfResults = 1;
+    r.body = JSON.stringify(parsed);
+    return r;
   },
 });
 const {decorator: searchInterfaceDecorator, play} =
   wrapInSearchInterface(engineConfig);
 
 const meta: Meta = {
-  component: 'atomic-result-image',
-  title: 'Atomic/ResultList/ResultImage',
-  id: 'atomic-result-image',
+  component: 'atomic-result-number',
+  title: 'Atomic/ResultList/ResultNumber',
+  id: 'atomic-result-number',
   render: renderComponent,
   decorators: [resultDecorator, searchInterfaceDecorator],
   parameters: {
@@ -31,9 +30,13 @@ const meta: Meta = {
 export default meta;
 type Story = StoryObj;
 
-export const FirstStory: Story = {
-  name: 'atomic-result-image',
+export const Default: Story = {
+  name: 'atomic-result-date',
   args: {
-    field: 'randomimage',
+    default: `
+      <span style="font-weight: bold; margin-right: 0.25rem;">File size:</span>
+      <atomic-format-unit unit="byte" unit-display="long"></atomic-format-unit>
+    `,
+    field: 'size',
   },
 };
