+++ packages/atomic/src/components/search/result-template-components/atomic-result-number/atomic-result-number.tsx
@@ -11,7 +11,7 @@ import {ResultContext} from '../result-template-decorators';
 /**
  * The `atomic-result-number` component renders the value of a number result field.
  *
- * The number can be formatted by adding a `atomic-format-number`, `atomic-format-currency` or `atomic-format-unit` component into this component.
+ * @slot default - The number can be formatted by adding a `atomic-format-number`, `atomic-format-currency` or `atomic-format-unit` component into this component.
  */
 @Component({
   tag: 'atomic-result-number',
