--- packages/atomic/src/components/search/result-template-components/atomic-result-timespan/atomic-result-timespan.stories.tsx
@@ -1,23 +1,0 @@
-import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'atomic-result-timespan',
-  {field: 'ytvideoduration', unit: 's'},
-  {
-    engineConfig: {
-      preprocessRequest: (r) => {
-        const request = JSON.parse(r.body!.toString());
-        request.cq = '@ytvideoduration';
-        request.fieldsToInclude = ['ytvideoduration'];
-        r.body = JSON.stringify(request);
-        return r;
-      },
-    },
-  }
-);
-export default {
-  ...defaultModuleExport,
-  title: 'Atomic/ResultList/ResultTimespan',
-  id: 'atomic-result-timespan',
-};
-export const Default = exportedStory;
