+++ packages/atomic/storybookUtils/result-wrapper.tsx
@@ -1,9 +1,5 @@
-import {
-  SearchEngineConfiguration,
-  getSampleSearchEngineConfiguration,
-} from '@coveo/headless';
-import {within} from '@storybook/test';
-import {Decorator, StoryContext} from '@storybook/web-components';
+import {SearchEngineConfiguration} from '@coveo/headless';
+import {Decorator} from '@storybook/web-components';
 import {html, render} from 'lit';
 import {unsafeHTML} from 'lit/directives/unsafe-html.js';
 import type * as _ from '../src/components.js';
@@ -25,10 +21,10 @@ const preprocessRequestForOneResult = (r: Request) => {
 };
 
 export const wrapInResult = (
-  config?: Partial<SearchEngineConfiguration>
+  engineConfig?: Partial<SearchEngineConfiguration>
 ): {
   decorator: Decorator;
-  play: (context: StoryContext) => Promise<void>;
+  engineConfig: Partial<SearchEngineConfiguration>;
 } => ({
   decorator: (story) => {
     const tempResultTemplate = document.createElement('div');
@@ -37,50 +33,34 @@ export const wrapInResult = (
       ...Array.from(tempResultTemplate.children)
     );
     return html`
-      <atomic-search-interface data-testid="root-interface">
-        <div style="position: relative; margin-top: 20px;">
-          <atomic-result-list
-            display="list"
-            density="normal"
-            image-size="icon"
-            style="border: 2px dashed black; padding:20px; position: relative;"
-          >
-            <atomic-result-template results>
-              ${unsafeHTML(
-                `<template>${tempResultTemplate.innerHTML}</template>`
-              )}
-            </atomic-result-template>
-          </atomic-result-list>
-          <div style="position: absolute; top: -20px; right: 0;">Template</div>
-        </div>
-        <style>
-          atomic-search-interface,
-          atomic-result-list {
-            max-width: 1024px;
-            display: block;
-            margin: auto;
-          }
-        </style>
-      </atomic-search-interface>
+      <div style="position: relative; margin-top: 20px;">
+        <atomic-result-list
+          display="list"
+          density="normal"
+          image-size="icon"
+          style="border: 2px dashed black; padding:20px; position: relative;"
+        >
+          <atomic-result-template results>
+            ${unsafeHTML(
+              `<template>${tempResultTemplate.innerHTML}</template>`
+            )}
+          </atomic-result-template>
+        </atomic-result-list>
+        <div style="position: absolute; top: -20px; right: 0;">Template</div>
+      </div>
+      <style>
+        atomic-search-interface,
+        atomic-result-list {
+          max-width: 1024px;
+          display: block;
+          margin: auto;
+        }
+      </style>
       <div style="hidden">${unsafeHTML(tempResultTemplate.innerHTML)}</div>
     `;
   },
-  play: async ({canvasElement, step}) => {
-    await customElements.whenDefined('atomic-search-interface');
-    const canvas = within(canvasElement);
-    const searchInterface =
-      await canvas.findByTestId<HTMLAtomicSearchInterfaceElement>(
-        'root-interface'
-      );
-    await step('Render the Search Interface', async () => {
-      await searchInterface!.initialize({
-        ...getSampleSearchEngineConfiguration(),
-        ...{preprocessRequest: preprocessRequestForOneResult},
-        ...config,
-      });
-    });
-    await step('Execute the first search', async () => {
-      await searchInterface!.executeFirstSearch();
-    });
+  engineConfig: {
+    preprocessRequest: preprocessRequestForOneResult,
+    ...engineConfig,
   },
 });
